requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
gunicorn>=21.2.0
sentence-transformers>=2.2.2
scikit-learn>=1.8.0
numpy>=1.25.0
joblib>=1.4.0
pandas>=2.2.0
matplotlib>=3.7.1
seaborn>=0.12.2
pydantic>=2.5.2
slowapi>=0.1.9
python-dotenv>=1.0.0
python-multipart>=0.0.6
